requirements.txt
langchain>=0.1.16
langgraph>=0.0.33
langchain-google-genai>=0.0.7
google-generativeai>=0.4.1
gradio>=4.19.2
python-dotenv>=1.0.1


